instadoc/requirements/devel.txt
# deployment
Fabric
Jinja2


# test runner
nose
django-nose


# test plugins
nosegrowl2


# python mocker
#fudge
